requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow>=10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==1.0.0
reportlab==4.0.7
